Refactor: remove PermissionsOverviewGlobal and PermissionsOverviewOrganizations components
